// ht-web-interface/src/models/family/FamilyEvent.ts
export enum FamilyEvent {
    Name = "family",
    Create = "family/CREATE",
    CreateSuccess = "family/CREATE_SUCCESS",
    CreateFailure = "family/CREATE_FAILURE",
    Edit = "family/EDIT",
    EditSuccess = "family/EDIT_SUCCESS",
    EditFailure = "family/EDIT_FAILURE",
    Reset = "family/RESET",

}